refactor: Tika classes to /Tika
