+++ elasticms-cli/src/Client/Audit/AuditManager.php
@@ -11,9 +11,9 @@ use App\CLI\Helper\AsyncResponse;
 use App\CLI\Helper\HtmlHelper;
 use App\CLI\Helper\LighthouseWrapper;
 use App\CLI\Helper\Pa11yWrapper;
-use App\CLI\Helper\TikaClient;
-use App\CLI\Helper\TikaMetaResponse;
-use App\CLI\Helper\TikaWrapper;
+use App\CLI\Helper\Tika\TikaClient;
+use App\CLI\Helper\Tika\TikaMetaResponse;
+use App\CLI\Helper\Tika\TikaWrapper;
 use EMS\CommonBundle\Common\Converter;
 use EMS\Helpers\Standard\Json;
 use Psr\Log\LoggerInterface;
