+++ elasticms-cli/src/Helper/Tika/TikaWrapper.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\CLI\Helper;
+namespace App\CLI\Helper\Tika;
 
+use App\CLI\Helper\ProcessWrapper;
 use EMS\Helpers\Standard\Text;
 use Psr\Http\Message\StreamInterface;
